requirements.txt
click==8.1.3
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
phonenumbers==8.12.49
psycopg2-binary==2.9.3
PyJWT==2.4.0
validate-docbr==1.10.0
Werkzeug==2.1.2
